requirements.txt
mindsdb
slack-bolt
tweepy
discord.py
textblob
python-dotenv